Updating to 2.7: Expand graph to fill screen.
This is a different method of making the BG graph fill screens compared to my attempt on v2.6. Instead of removing the scroll view altogether, I've instead used `android:fillViewport` on the `ScrollView` which lets the view group stretch all the way down to the buttons. For the graph(s) layout, I set the `bggraph` layout height to `wrap_content` because otherwise the graph wouldn't expand due to it being set to 200dp by default. Lastly, I had to remove a line of code in `OverviewFragment.kt` that also set the `bggraph` layout height (although this value was determined by the skin you're using on AAPS).