Fix insulin changes on MDT pumps being auto-logged as cannula changes.
When replacing the insulin reservoir independently of the cannula, I find that the tubing should be re-primed or replaced altogether to ensure there are no air bubbles from when the reservoir was disconnected from the tubing. When the tubing is primed, its history entry is the same type as a fixed-prime so any time the reservoir was replaced AAPS would log a site change which reset autosens unnecessarily. This commit aims to fix that by checking the `FixedAmount` data entry of `Prime` history entries to detect whether the prime was a manual-prime or fixed-prime and skip logging a site change event if it is the latter.

In order to automatically log insulin reservoir replacements, this commit also monitors `Rewind` history entries and logs an `InsulinChange` event for them (similar to the prime event monitor).